Mahindra Santacruz Premium Residences In Mumbai
Mahindra Santacruz is a new launch residential project by Mahindra Lifespaces coming up at Santacruz West, Mumbai. The property is efficiently planned with the condition of workmanship highlighting the features that offer you an ideal mix of comfort and solace with a bit of a higher standard of living for you with its better subtleties.
The property extends over wide land parcels and provides a composite hub of 2 BHK and 3 BHK apartments that are majestic and spectacular fully loaded with all the hi-tech specifications and provide you with a truly blissful living experience. The interiors are beautifully done with top-class quality fittings. The property is well-supported with the aid of easy payment plans.
Mahindra Lifespaces Santacruz Mumbai comprises a horde of lifestyle amenities that make it a one-of-its-kind unique project. The palatial services offered here are a senior citizen corner, tennis court, jogging track, party lawn, clubhouse and gym, game club, crèche, infinity pool with toddlers’ pool, mini-theatre, library, yoga room and loads of different posh activities. Its ultra-contemporary security features consist of multi-layer security, a fire safety management system, smoke detectors, a panic alarm, a video door phone, etc.
Mahindra Lifespaces Santacruz is coming up in the famous locality of Santacruz (W), Mumbai and has been built up taking care of every tiny detail and with s sole purpose to offer you a residing venture that is above the rest. Everything essential like schools, colleges, hospitals, markets, banks & ATMs are located just round the corner. It is 18-10 minutes away from Mumbai airport and at a distance of 4-5 km away from the Hi-Life Mall and Link Square Mall.
Being placed in a well-developed residential vicinity with good infrastructure, this Mahindra Santacruz Property is worth funding and with all its state-of-art features and amenities, it is sure to be the most successful project in the coming years making it the most profitable investment too for both buyers and investors.
To know more - https://bit.ly/3YAoh1j